AMP up recipe-app.com
This CL attempts to AMP up recipe-app.com in an effort to show developers
how they can take their existing sites and make them AMP'ed.

Since recipe-app.com is not a mobile-friendly site currently
(the whole website just opens up in the browser), the hope is to make
it responsive in the future.

There are two ways to implement AMP:
1) Have regular html pages, and separate AMP'ed up pages for all those,
which is the approach taken in this CL
2) Have only AMP pages across the site. This is acheivable in the future,
if we can make the site mobile-friendly and apply consistent
CSS across desktop/phone/other devices where content scales to the size of screen

What does that mean for this CL?
We determine if an AMP page is requested by looking at the URL of the page. If it has
"amp/" at the beginning, we know they requested an AMP page (amp-recipe.html)
 and take them to those recipe files instead.
Otherwise, a canonical page is requested and they go to the non-AMP recipes pages (recipe.html)
When we can get (2) from above, we can consolidate and use a single template.

Some specific AMP rules that needed to be followed for this CL to pass AMP validation:
- No external CSS files. All the CSS had to be consolidated and put into
the header. only one <style> tag is allowed.
- All images have to be marked as amp-img and are responsive based on phone screen size
- Forms needed target actions
- CSS styling that was specific to desktop cases or "banned" in AMP were removed (things like
the universal selector ie. *)
- All AMP pages link to their canonical counterparts, and (vice versa) all canonical pages link to their AMP pages.

Tests:
1. AMP pages are valid and validated through the AMPed validator: https://screenshot.googleplex.com/Purvds4RMrw.png
2. New tests written for the recipe_template_for_url function pass: https://screenshot.googleplex.com/Qd0kxmPH7mK.png
3. Structured data test tool also shows AMP icon for these AMP pages now: https://screenshot.googleplex.com/OiqMvxxsGSn.png
Change-Id: I977fbc37c22f9163200692386a15ab77393979a3